+++ scripts/macros/monsterFeatures/zombie/undeadFortitude.js
@@ -1,4 +1,5 @@
 import {chris} from '../../../helperFunctions.js';
+import {queue} from '../../../queue.js';
 export async function undeadFortitude(targetToken, {workflow, ditem}) {
     if (ditem.newHP != 0 || ditem.oldHP === 0) return;
     let targetActor = targetToken.actor;
@@ -8,6 +9,8 @@ export async function undeadFortitude(targetToken, {workflow, ditem}) {
     if (!effect) return;
     let originItem = await fromUuid(effect.origin);
     if (!originItem) return;
+    let queueSetup = await queue.setup(workflow.uuid, 'undeadFortitude', 389);
+    if (!queueSetup) return;
     let featureData = duplicate(originItem.toObject());
     let damageDealt = ditem.appliedDamage;
     featureData.system.save.dc = damageDealt + featureData.system.save.dc;
@@ -22,7 +25,11 @@ export async function undeadFortitude(targetToken, {workflow, ditem}) {
         'consumeSlot': false,
     };
     let featureWorkflow = await MidiQOL.completeItemUse(feature, {}, options);
-    if (featureWorkflow.failedSaves.size === 1) return;
+    if (featureWorkflow.failedSaves.size === 1) {
+        queue.remove(workflow.uuid);
+        return;
+    }
     ditem.newHP = 1;
     ditem.hpDamage = Math.abs(ditem.newHP - ditem.oldHP);
+    queue.remove(workflow.uuid);
 }